bot/requirements.txt
# bot/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
discord.py
python-dotenv
requests
PyGithub
qdrant-client
sentence-transformers
watchfiles # <-- DODAJ TĘ LINIĘ